Fix unary minus for lessons 02 and 03
Fix both compilation and interpretation of unary minus operator. The
compiler forgot to generate the bytecode for the operand, and the VM
erroneously handled the + operator like the - operator.

Note that this commit only fixes up lessons 02 and 03, as lesson 04
does not even run inside the Lua interpreter, so I won't try to fix it
anyway...