Add integration tests for lessons 01 to 03
Include a TEST_MODE in the VM which will take the bytecode as a hex string
from the command line instead of having it hardcoded in a header file. Add
integration test lua files that call out to the compiled TEST_MODE VM to
test the compiled bytecode (This was how I found the bugs in the unary
minus implementation).

Note that this commit only adds it for lessons 01 to 03, as lesson 04 does
not even run inside the Lua interpreter, so I won't try to fix it
anyway...